examples/requirement.txt
facenet-pytorch
numpy>=1.24.0,<2.0.0
requests>=2.0.0,<3.0.0
torch>=2.2.0,<2.3.0
torchvision>=0.17.0,<0.18.0
Pillow>=10.2.0,<10.3.0
opencv-python>=4.9.0
scipy>=1.10.0,<2.0.0
tqdm>=4.0.0,<5.0.0
pandas>=2.0.0,<3.0.0
coverage>=7.0.0,<8.0.0
codecov>=2.0.0,<3.0.0
jupyter>=1.0.0
tensorboard>=2.0.0,<3.0.0
pytest-runner
docutils==0.16.0
markdown>=3.4.0
myst-parser
opencv-python
-e git+https://github.com/open-mmlab/pytorch_sphinx_theme.git#egg=pytorch_sphinx_theme
sphinx==4.0.2
sphinx-copybutton
sphinx_markdown_tables>=0.0.16
torch
urllib3<2.0.0
ninja
psutil
addict
mmengine>=0.3.0
numpy
packaging
Pillow
pyyaml
regex;sys_platform=='win32'
yapf
coverage
lmdb
onnx
onnxoptimizer
onnxruntime
pytest
PyTurboJPEG
scipy
tifffile

